+++ src/components/core/layout/Flexbox/Flexbox-styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Flexbox = styled('div')({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+  '& > * + *': {
+    marginLeft: '30rem',
+  },
+  '@media (max-width: 768px)': {
+    '& > * + *': {
+      marginLeft: '15rem',
+    },
+  },
+});